// angular-14-client/src/app/models/Cliente.model.ts
export class Cliente {
  id_cliente?: number;
  nombres?: string;
  apellidos?: string;
  nro_documento?: string;
  email?: string;
  //create_at?: Date;
  //ventas?:Venta[];

}

/* export class Venta { 
  id_venta?: number;
  fecha?: Date;
  serie?: string;
  numero?: string; 
  total?: number; 
} */